Move gravatar_url to libs
