+++ vilya/templates/blame.html
@@ -10,7 +10,7 @@
 ${parent.body()}
 
 <%!
-    from vilya.libs.text import gravatar_url
+    from vilya.libs.gravatar import gravatar_url
     from urllib import urlencode
 %>
